# emas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  handler.SchedulingSettingsResponse:
    properties:
      auto_reschedule_on_event:
        type: boolean
      deviation_penalty_weight:
        type: number
      lateness_weight:
        type: number
      lock_in_window_minutes:
        type: integer
      objective:
        type: string
      public_holidays:
        items:
          type: string
        type: array
      setup_weight:
        type: number
      slack_weight:
        type: number
      split_strategy:
        type: string
      updated_at:
        type: string
      work_days:
        type: string
      work_end_time:
        type: string
      work_start_time:
        type: string
    type: object
...
  handler.UpdateSchedulingSettingsRequest:
    properties:
      auto_reschedule_on_event:
        type: boolean
      deviation_penalty_weight:
        type: number
      lateness_weight:
        type: number
      lock_in_window_minutes:
        type: integer
      objective:
        type: string
      public_holidays:
        items:
          type: string
        type: array
      setup_weight:
        type: number
      slack_weight:
        type: number
      split_strategy:
        type: string
      work_days:
        type: string
      work_end_time:
        type: string
      work_start_time:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Factory operations API.
  title: eMas Factory API
  version: "1.0"
paths:
...
  /scheduling/is-time-before:
    get:
      consumes:
      - application/json
      description: Is time before
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Is time before
      tags:
      - scheduling
  /scheduling/is-valid-iso-date:
    get:
      consumes:
      - application/json
      description: Is valid ISO date
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Is valid ISO date
      tags:
      - scheduling
...
  /scheduling/settings:
    put:
      consumes:
      - application/json
      description: Update scheduling settings
      parameters:
      - description: Update Scheduling Settings Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSchedulingSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SchedulingSettingsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update scheduling settings
      tags:
      - scheduling
...
  /scheduling/validate-work-days:
    get:
      consumes:
      - application/json
      description: Validate work days
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data: {}
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Validate work days
      tags:
      - scheduling
...
swagger: "2.0"